+++ lib/BicBucStriim/calibre_thing.php
@@ -8,16 +8,18 @@
  *
  */
 
- /**
-  * @property mixed $ownLink
-  * @property mixed $ownNote
-  * @property mixed $ownArtefact
-  * @property mixed $ctype
-  * @property mixed $cid
-  * @property mixed $cname
-  * @property mixed $refctr
-  */
-class Model_CalibreThing extends RedBean_SimpleModel
+/**
+ * RedBeanPHP FUSE object for 'calibrething' with one-to-many relation of author with link, note and artefact
+ * See https://www.redbeanphp.com/index.php?p=/one_to_many
+ * @property mixed $ownLinkList
+ * @property mixed $ownNoteList
+ * @property mixed $ownArtefactList
+ * @property mixed $ctype
+ * @property mixed $cid
+ * @property mixed $cname
+ * @property mixed $refctr
+ */
+class Model_CalibreThing extends \RedBeanPHP\SimpleModel
 {
     /**
      * Return author links releated to this Calibre entitiy.
@@ -25,18 +27,22 @@ class Model_CalibreThing extends RedBean_SimpleModel
      */
     public function getAuthorLinks()
     {
-        return array_values(array_filter($this->ownLink ?? [], function ($link) {
+        // Note: we cannot use $this->ownLinkList ?? [] directly here due to lazy loading
+        $links = $this->ownLinkList;
+        return array_values(array_filter($links ?? [], function ($link) {
             return($link->ltype == DataConstants::AUTHOR_LINK);
         }));
     }
 
     /**
      * Return the author note text related to this Calibre entitiy.
-     * @return ?RedBean_SimpleModel 	text or null
+     * @return ?\RedBeanPHP\SimpleModel 	text or null
      */
     public function getAuthorNote()
     {
-        $notes = array_values(array_filter($this->ownNote ?? [], function ($note) {
+        // Note: we cannot use $this->ownNoteList ?? [] directly here due to lazy loading
+        $notes = $this->ownNoteList;
+        $notes = array_values(array_filter($notes ?? [], function ($note) {
             return($note->ntype == DataConstants::AUTHOR_NOTE);
         }));
         if (empty($notes)) {
@@ -48,11 +54,13 @@ class Model_CalibreThing extends RedBean_SimpleModel
 
     /**
      * Return the author thumbnail file related to this Calibre entitiy.
-     * @return ?RedBean_SimpleModel 	Path to thumbnail file or null
+     * @return ?\RedBeanPHP\SimpleModel 	Path to thumbnail file or null
      */
     public function getAuthorThumbnail()
     {
-        $artefacts = array_values(array_filter($this->ownArtefact ?? [], function ($artefact) {
+        // Note: we cannot use $this->ownArtefactList ?? [] directly here due to lazy loading
+        $artefacts = $this->ownArtefactList;
+        $artefacts = array_values(array_filter($artefacts ?? [], function ($artefact) {
             return($artefact->atype == DataConstants::AUTHOR_THUMBNAIL_ARTEFACT);
         }));
         if (empty($artefacts)) {